Fixed some bugs, added support for an AdaFruit HT16K33 4 Digit 7 Segment LED Display to display group state and info.

+++ input_gpio.h
@@ -30,8 +30,8 @@ class InputGpio {
   volatile uint8_t processed_event_counter;
   volatile ProcessedEvent processed_event_queue[MAX_EVENT_QUEUE_SIZE];
   InputProcessedEvent last_event;
-  int last_track;
-  int last_group;
+  uint16_t last_track;
+  uint8_t last_group;
   bool last_event_for_track;
 
   // Methods
